+++ server/src/api/chat/chatRoom.mutation.ts
@@ -1,0 +1,61 @@
+import {
+  MutationResolvers,
+  MutationCreateChatRoomArgs,
+  Chat,
+} from '../../types';
+import { requestDB } from '../../utils/requestDB';
+import createDBError from '../../errors/createDBError';
+import isAuthenticated from '../../utils/isAuthenticated';
+import {
+  CHECK_CHAT_ROOM_QUERY,
+  CREATE_CHAT_ROOM_QUERY
+} from '../../schema/chat/chatQuery';
+import { parseResultRecords } from '../../utils/parseData';
+import { createChatAndPublish, publishToMessageTab } from './chat.pubsub';
+
+const Mutation: MutationResolvers = {
+  createChatRoom: async (
+    _,
+    { userEmail, content }: MutationCreateChatRoomArgs,
+    { req, pubsub }
+  ): Promise<number> => {
+    isAuthenticated(req);
+    const { email } = req;
+    try {
+      const checkChatRoomResult = await requestDB(CHECK_CHAT_ROOM_QUERY, {
+        chatMemberEmail1: email,
+        chatMemberEmail2: userEmail
+      });
+
+      if (checkChatRoomResult.length) {
+        const [{ chatRoomId }] = parseResultRecords(checkChatRoomResult);
+        await createChatAndPublish({
+          email,
+          content,
+          chatRoomId: Number(chatRoomId),
+          pubsub
+        });
+        return chatRoomId;
+      }
+
+      const result = await requestDB(CREATE_CHAT_ROOM_QUERY, {
+        from: email,
+        to: userEmail,
+        content
+      });
+      const { chats }: { chats: Chat[] } = parseResultRecords(result)[0];
+      const chat = chats[0];
+      publishToMessageTab({
+        pubsub,
+        chatRoomId: chat.chatRoomId,
+        chat
+      });
+      return chat.chatRoomId;
+    } catch (error) {
+      const DBError = createDBError(error);
+      throw new DBError();
+    }
+  }
+};
+
+export default Mutation;
