+++ server/src/api/chat/filterFunction.ts
@@ -1,6 +1,6 @@
 import { requestDB } from '../../utils/requestDB';
 import { parseResultRecords } from '../../utils/parseData';
-import { GET_USERS_ON_CHAT_ROOM_QUERY } from 'src/schema/chat/chatQuery';
+import { GET_USERS_ON_CHAT_ROOM_QUERY } from '../../schema/chat/chatQuery';
 
 export const filterChatRoomUser = async ({ chatRoomId, email }) => {
   const result = await requestDB(GET_USERS_ON_CHAT_ROOM_QUERY, {
